+++ src/pytorch_lightning/plugins/precision/deepspeed.py
@@ -11,14 +11,14 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from typing import Any, Callable, Optional, TYPE_CHECKING, Union
+from typing import Any, Callable, cast, Optional, TYPE_CHECKING, Union
 
 from lightning_utilities.core.imports import RequirementCache
 from torch import Tensor
 from torch.optim import LBFGS, Optimizer
+from typing_extensions import get_args, Literal
 
 import pytorch_lightning as pl
-from lightning_fabric.utilities.enums import PrecisionType
 from lightning_fabric.utilities.types import Steppable
 from pytorch_lightning.plugins.precision.apex_amp import _APEX_AVAILABLE
 from pytorch_lightning.plugins.precision.precision_plugin import PrecisionPlugin
@@ -33,26 +33,33 @@ if TYPE_CHECKING and _DEEPSPEED_AVAILABLE:
 
 warning_cache = WarningCache()
 
+_PRECISION_INPUT_INT = Literal[32, 16]
+_PRECISION_INPUT_STR = Literal["32", "16", "bf16"]
+_PRECISION_INPUT = Union[_PRECISION_INPUT_INT, _PRECISION_INPUT_STR]
+
 
 class DeepSpeedPrecisionPlugin(PrecisionPlugin):
     """Precision plugin for DeepSpeed integration.
 
     Args:
-        precision: Double precision (64), full precision (32), half precision (16) or bfloat16 precision (bf16).
+        precision: Full precision (32), half precision (16) or bfloat16 precision (bf16).
     Raises:
         ValueError:
             If unsupported ``precision`` is provided.
     """
 
     def __init__(
-        self, precision: Union[str, int], amp_type: Optional[str] = None, amp_level: Optional[str] = None
+        self,
+        precision: Literal["32", 32, "16", 16, "bf16"],
+        amp_type: Optional[str] = None,
+        amp_level: Optional[str] = None,
     ) -> None:
         if amp_type == "apex":
-            # TODO: remove in v1.10.0
+            # TODO: remove in v2.0.0
             rank_zero_deprecation(
                 "The NVIDIA/apex AMP implementation has been deprecated upstream. Consequently, its integration inside"
                 " PyTorch Lightning has been deprecated in v1.9.0. Support for using it through the DeepSpeed"
-                " implementation will be removed in v1.10.0."
+                " implementation will be removed in v2.0.0."
             )
             if not _APEX_AVAILABLE:
                 raise MisconfigurationException(
@@ -70,18 +77,16 @@ class DeepSpeedPrecisionPlugin(PrecisionPlugin):
         else:
             rank_zero_deprecation(
                 f"Passing `{type(self).__name__}(amp_type={amp_type!r})` been deprecated in v1.9.0 and will be removed"
-                f" in v1.10.0. This argument is no longer necessary."
+                f" in v2.0.0. This argument is no longer necessary."
             )
 
-        supported_precision = (PrecisionType.HALF, PrecisionType.FLOAT, PrecisionType.BFLOAT)
+        supported_precision = get_args(_PRECISION_INPUT_STR) + get_args(_PRECISION_INPUT_INT)
         if precision not in supported_precision:
             raise ValueError(
                 f"`Trainer(strategy='deepspeed', precision={precision!r})` is not supported."
-                f" `precision` must be one of: {(x.value for x in supported_precision)}."
+                f" `precision` must be one of: {supported_precision}."
             )
-
-        super().__init__()
-        self.precision = precision
+        self.precision = cast(_PRECISION_INPUT_STR, str(precision))
         self.amp_type = amp_type
         self.amp_level = amp_level