+++ tests/tests_pytorch/trainer/test_trainer.py
@@ -1605,7 +1605,7 @@ def test_setup_hook_move_to_device_correctly(tmpdir, accelerator):
             output = self.layer(batch)
             # will crash if not moved to correct device
             output = self.new_layer(output)
-            loss = self.loss(batch, output)
+            loss = self.loss(output)
             return {"loss": loss}
 
     # fake data
@@ -1745,8 +1745,7 @@ def test_model_in_correct_mode_during_stages(tmpdir, strategy, devices):
 
 class TestDummyModelForCheckpoint(BoringModel):
     def validation_step(self, batch, batch_idx):
-        output = self.layer(batch)
-        loss = self.loss(batch, output)
+        loss = self.step(batch)
         self.log("x", loss)
 
     def validation_epoch_end(self, outputs) -> None:
@@ -1922,7 +1921,7 @@ class ExceptionCounter(Callback):
         self.exceptions += 1
 
 
-@pytest.mark.parametrize("strategy", [None, pytest.param("ddp_spawn", marks=RunIf(skip_windows=True))])
+@pytest.mark.parametrize("strategy", [None, pytest.param("ddp_spawn", marks=RunIf(skip_windows=True, mps=False))])
 def test_error_handling_all_stages(tmpdir, strategy):
     model = TrainerStagesErrorsModel()
     counter = ExceptionCounter()
@@ -2018,9 +2017,11 @@ def test_detect_anomaly_nan(tmpdir):
     ["trainer_kwargs", "strategy_cls", "strategy_name", "accelerator_cls", "devices"],
     [
         ({"strategy": None}, SingleDeviceStrategy, "single_device", CPUAccelerator, 1),
-        ({"strategy": "dp"}, DDPStrategy, "ddp", CPUAccelerator, 1),
-        ({"strategy": "ddp"}, DDPStrategy, "ddp", CPUAccelerator, 1),
-        ({"strategy": "ddp", "num_nodes": 2}, DDPStrategy, "ddp", CPUAccelerator, 1),
+        pytest.param({"strategy": "dp"}, DDPStrategy, "ddp", CPUAccelerator, 1, marks=RunIf(mps=False)),
+        pytest.param({"strategy": "ddp"}, DDPStrategy, "ddp", CPUAccelerator, 1, marks=RunIf(mps=False)),
+        pytest.param(
+            {"strategy": "ddp", "num_nodes": 2}, DDPStrategy, "ddp", CPUAccelerator, 1, marks=RunIf(mps=False)
+        ),
         (
             {"strategy": None, "accelerator": "cuda", "devices": 1},
             SingleDeviceStrategy,
@@ -2076,7 +2077,7 @@ def test_detect_anomaly_nan(tmpdir):
             CUDAAccelerator,
             2,
         ),
-        ({"strategy": DDPStrategy()}, DDPStrategy, "ddp", CPUAccelerator, 1),
+        pytest.param({"strategy": DDPStrategy()}, DDPStrategy, "ddp", CPUAccelerator, 1, marks=RunIf(mps=False)),
         ({"strategy": DDPStrategy(), "accelerator": "cuda", "devices": 2}, DDPStrategy, "ddp", CUDAAccelerator, 2),
         (
             {"strategy": DataParallelStrategy(), "accelerator": "cuda", "devices": 2},